More progress on matches.
Fixed a bug where the starting point would jump and create crazy
matches!

Looks like the regex is broken, need to fix that.  Also I need to have
the stack let the user know if that is the current statement, it matters
heavily when figuring out what to do with a line of code if the current
line is setting the new stack or just within the current stack.